refactor(web): split AdminSettings cards into pages/adminSettings/ (#960 M2)
Pure motion: move the ten named AdminSettings card components, each with the
private helpers only it uses, out of the AdminSettings.tsx shell into one file
each under new web/src/pages/adminSettings/. The shell keeps AdminSettings()
and its three inline cards (vault-migration, forge-labels, sso) plus
oidcStatusChip and clientValidate, and shrinks 2506 -> 431 lines. Each card
reaches the same ../../lib/api module the tests mock, so AdminSettings.test.tsx
and AdminSettingsUpdates.test.tsx pass unmodified. Doc repoints in
ARCHITECTURE.md and specs/ai.md.

Part of PRD #960 (epic #915, finding W5).
